requirements.txt
Flask==2.0.2
gunicorn==20.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
